# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "service stores \u0026 provides feed for users",
        "title": "makala Feed Service",
        "contact": {
            "name": "Yerassyl Danay"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/verification",
    "paths": {
        "/api/makala/v1/feed": {
            "get": {
                "description": "this API lets a user fetch feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "fetches feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of elements",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "author name for ad index caching",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "time, when a user started fetching feed",
                        "name": "started_fetching_at_unix_nano_utc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetFeedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrMsg"
                        }
                    }
                }
            }
        },
        "/api/makala/v1/post": {
            "post": {
                "description": "creates a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "creates a post",
                "parameters": [
                    {
                        "description": "post info",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/poststore.FeedPost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrMsg": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
        "model.GetFeedResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/poststore.FeedPost"
                    }
                },
                "started_fetching_at_unix_nano_utc": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.PostRequest": {
            "type": "object",
            "required": [
                "author"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "t2_11qnzrqv"
                },
                "content": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "nsfw": {
                    "type": "boolean"
                },
                "promoted": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "submakala": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "poststore.FeedPost": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "nsfw": {
                    "type": "boolean"
                },
                "promoted": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "submakala": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}