# provider/poststore/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.15.0

package poststore

import ()

type FeedPost struct {
	ID        int64   `json:"id"`
	Title     string  `json:"title"`
	Author    string  `json:"author"`
	Link      *string `json:"link"`
	Submakala string  `json:"submakala"`
	Content   *string `json:"content"`
	Score     float64 `json:"score"`
	Promoted  bool    `json:"promoted"`
	Nsfw      bool    `json:"nsfw"`
}